requirements.txt
lockfile==0.12.2
nose==1.3.7
pika==0.11.1
python-daemon==2.1.2
python-dateutil==2.6.1
rednose==1.2.3
